some unnecessary code deleted
