+++ Movie/Scene/MovieDetails/MovieDetailsModels.swift
@@ -14,9 +14,7 @@ enum MovieDetails {
   enum Models {
    
     struct MovieDetailsResponse: Codable {
-        let adult: Bool?
         let backdropPath: String?
-        let belongsToCollection: BelongsToCollection?
         let budget: Int?
         let genres: [Genre]?
         let homepage: String?
@@ -24,32 +22,25 @@ enum MovieDetails {
         let imdbID, originalLanguage, originalTitle, overview: String?
         let popularity: Double?
         let posterPath: String?
-        let productionCompanies: [ProductionCompany]?
-        let productionCountries: [ProductionCountry]?
         let releaseDate: String?
-        let revenue, runtime: Int?
+        let revenue: Int?
         let spokenLanguages: [SpokenLanguage]?
-        let status, tagline, title: String?
-        let video: Bool?
+        let status, title: String?
         let voteAverage: Double?
         let voteCount: Int?
 
         enum CodingKeys: String, CodingKey {
-            case adult
             case backdropPath = "backdrop_path"
-            case belongsToCollection = "belongs_to_collection"
             case budget, genres, homepage, id
             case imdbID = "imdb_id"
             case originalLanguage = "original_language"
             case originalTitle = "original_title"
             case overview, popularity
             case posterPath = "poster_path"
-            case productionCompanies = "production_companies"
-            case productionCountries = "production_countries"
             case releaseDate = "release_date"
-            case revenue, runtime
+            case revenue
             case spokenLanguages = "spoken_languages"
-            case status, tagline, title, video
+            case status, title
             case voteAverage = "vote_average"
             case voteCount = "vote_count"
         }
@@ -73,29 +64,6 @@ enum MovieDetails {
         let name: String?
     }
 
-    // MARK: - ProductionCompany
-    struct ProductionCompany: Codable {
-        let id: Int?
-        let logoPath, name, originCountry: String?
-
-        enum CodingKeys: String, CodingKey {
-            case id
-            case logoPath = "logo_path"
-            case name
-            case originCountry = "origin_country"
-        }
-    }
-
-    // MARK: - ProductionCountry
-    struct ProductionCountry: Codable {
-        let iso3166_1, name: String?
-
-        enum CodingKeys: String, CodingKey {
-            case iso3166_1 = "iso_3166_1"
-            case name
-        }
-    }
-
     // MARK: - SpokenLanguage
     struct SpokenLanguage: Codable {
         let englishName, iso639_1, name: String?
@@ -107,7 +75,6 @@ enum MovieDetails {
         }
     }
 
-   
       struct ViewModel {
         let overView: String
         let image: Data?
