+++ Movie/Scene/MovieDetails/MovieDetailsViewController.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 protocol MovieDetailsViewControllerDelegate {
-
     func viewDidload()
 }
 
@@ -53,17 +52,6 @@ class MovieDetailsViewController: UIViewController {
     presenter.viewController = viewController
     router.viewController = viewController
     router.dataStore = interactor
-  }
-  
-  // MARK: Routing
-  override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-  
-      if let scene = segue.identifier {
-      let selector = NSSelectorFromString("routeTo\(scene)WithSegue:")
-      if let router = router, router.responds(to: selector) {
-        router.perform(selector, with: segue)
-      }
-    }
   }
   
   // MARK: View lifecycle
