+++ Movie/Scene/SearchMovies/SearchMoviesWorker.swift
@@ -3,15 +3,15 @@
 //  Movie
 //
 //  Created by javad faghih on 9/6/22.
-//  Copyright (c) 2022 ___ORGANIZATIONNAME___. All rights reserved.
+//  Copyright (c) 2022 Javad Faghih. All rights reserved.
 //
 
 import UIKit
 import Alamofire
 
-class SearchMoviesWorker { //_ items: SearchMovieResponseModel?,_ error: NetworkErrorType?
+class SearchMoviesWorker {
     
-    func requestMovie(with word: String, page: Int = 1, complation: @escaping(_ result: Swift.Result<SearchMovieResponseModel, NetworkErrorType>) -> Void) {
+    func requestMovie(with word: String, page: Int = 1, complation: @escaping(_ result: Swift.Result<SearchMovies.Models.SearchMovieResponseModel, NetworkErrorType>) -> Void) {
         
         let baseURL = NetworkRequirment.baseURL.rawValue
         let searchMoviewEndPoint = Search.searchMovie.rawValue
@@ -27,7 +27,7 @@ class SearchMoviesWorker { //_ items: SearchMovieResponseModel?,_ error: Network
             
             case .success(let data):
                 
-                if  let model = try? JSONDecoder().decode(SearchMovieResponseModel.self, from: data) {
+                if  let model = try? JSONDecoder().decode(SearchMovies.Models.SearchMovieResponseModel.self, from: data) {
                     
                     if model.results.isEmpty {
                         complation(.failure(.noData))
