+++ Movie/Scene/Settings/Cell/LanguageTableViewCell.swift
@@ -17,12 +17,6 @@ class LanguageTableViewCell: MasterTableViewCell {
         super.awakeFromNib()
         // Initialization code
     }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
     
     func configure(title: String) {
         self.titleLabel.text = title
