+++ Movie/Scene/Settings/SettingsInteractor.swift
@@ -10,11 +10,9 @@ import UIKit
 
 protocol SettingsinteractorDelegate {
 
-    func presentSomething(response: Settings.Models.Response)
 }
 
 protocol SettingsDataStore {
-  //var name: String { get set }
 }
 
 typealias SettingsInteractorInput = SettingsViewControllerDelegate
@@ -22,8 +20,6 @@ typealias SettingsInteractorInput = SettingsViewControllerDelegate
 class SettingsInteractor: SettingsInteractorInput, SettingsDataStore {
 
     var presenter: SettingsinteractorDelegate?
-    var worker: SettingsWorker?
-  //var name: String = ""
   
   // MARK: Do something
     func viewDidload() {
