--- Movie/Scene/Settings/SettingsModels.swift
@@ -1,30 +1,0 @@
-//
-//  SettingsModels.swift
-//  Movie
-//
-//  Created by javad faghih on 9/10/22.
-//  Copyright (c) 2022 Javad Faghih. All rights reserved.
-
-
-import UIKit
-
-enum Settings {
- 
-    // MARK: Use cases
-  enum Models {
-   
-      struct Request {
-   
-    
-    }
-      struct Response {
-   
-    
-    }
-   
-      struct ViewModel {
-   
-      }
-      
-  }
-}
