--- Movie/Scene/Settings/SettingsWorker.swift
@@ -1,13 +1,0 @@
-//
-//  SettingsWorker.swift
-//  Movie
-//
-//  Created by javad faghih on 9/10/22.
-//  Copyright (c) 2022 ___ORGANIZATIONNAME___. All rights reserved.
-//
-
-import UIKit
-
-class SettingsWorker {
-  
-}
